ENH: move check topo out of main
